(Dashboard) Refactor utility functions, Make UI consistent
